// src/utils/interfaces.ts
export interface RatingProduct {
    rate: number;
    count: number;
}

export interface Product {
    id: number;
    title: string;
    price: number;
    description: string;
    category: string;
    image: string;
    rating: RatingProduct;
    ObjectBy?: string
}

export interface DataType {
    id: number;
    title: string;
    price: number;
    description: string;
    category: string;
    image: string;
    rating: RatingProduct;
    ObjectBy?: string
}